requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
tensorflow>=2.6.0
opencv-python>=4.5.3.56
numpy>=1.19.5
pymongo>=4.0.0
gridfs>=0.6.0
python-dotenv>=0.19.0
Pillow>=9.0.0